abstract and test sorted buf
