+++ polars/polars-arrow/src/kernels/rolling/mod.rs
@@ -1,6 +1,20 @@
 pub mod no_nulls;
 pub mod nulls;
+pub mod quantile_no_nulls;
+mod window;
+
+use crate::data_types::IsFloat;
+use crate::prelude::*;
+use crate::utils::CustomIterTools;
+use arrow::array::{ArrayRef, PrimitiveArray};
 use arrow::bitmap::{Bitmap, MutableBitmap};
+use arrow::types::NativeType;
+use num::ToPrimitive;
+use num::{NumCast, Zero};
+use std::cmp::Ordering;
+use std::ops::{Add, Div, Mul, Sub};
+use std::sync::Arc;
+use window::*;
 
 type Start = usize;
 type End = usize;
@@ -56,5 +70,44 @@ where
         Some(validity)
     } else {
         None
+    }
+}
+fn compare_fn<T>(a: &T, b: &T) -> Ordering
+where
+    T: PartialOrd + IsFloat + NativeType,
+{
+    if T::is_float() {
+        match (a.is_nan(), b.is_nan()) {
+            // safety: we checked nans
+            (false, false) => unsafe { a.partial_cmp(b).unwrap_unchecked() },
+            (true, true) => Ordering::Equal,
+            (true, false) => Ordering::Greater,
+            (false, true) => Ordering::Less,
+        }
+    } else {
+        // Safety:
+        // all integers are Ord
+        unsafe { a.partial_cmp(b).unwrap_unchecked() }
+    }
+}
+
+pub(super) fn sort_buf<T>(buf: &mut [T])
+where
+    T: IsFloat + NativeType + PartialOrd,
+{
+    if T::is_float() {
+        buf.sort_by(|a, b| {
+            match (a.is_nan(), b.is_nan()) {
+                // safety: we checked nans
+                (false, false) => unsafe { a.partial_cmp(b).unwrap_unchecked() },
+                (true, true) => Ordering::Equal,
+                (true, false) => Ordering::Greater,
+                (false, true) => Ordering::Less,
+            }
+        });
+    } else {
+        // Safety:
+        // all integers are Ord
+        unsafe { buf.sort_by(|a, b| a.partial_cmp(b).unwrap_unchecked()) };
     }
 }
